requirements.txt
absl-py==1.0.0
antlr4-python3-runtime==4.8
anyio==3.5.0
appdirs==1.4.4
asgiref==3.5.0
black==21.4b2
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.0
cloudpickle==2.0.0
cycler==0.11.0
-e git+https://github.com/facebookresearch/detectron2.git@6886f85baee349556749680ae8c85cdba1782d8e#egg=detectron2
fastapi==0.75.0
fonttools==4.31.2
future==0.18.2
fvcore==0.1.5.post20220305
google-auth==2.6.2
google-auth-oauthlib==0.4.6
grpcio==1.44.0
gunicorn==20.1.0
h11==0.13.0
hydra-core==1.1.1
idna==3.3
iopath==0.1.9
kiwisolver==1.4.2
Markdown==3.3.6
matplotlib==3.5.1
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.3
oauthlib==3.2.0
omegaconf==2.1.1
opencv-python==4.5.5.64
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
portalocker==2.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pydantic==1.9.0
pydot==1.4.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-multipart==0.0.5
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
toml==0.10.2
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.63.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
Werkzeug==2.1.0
yacs==0.1.8
